Docker

+++ configure-system.sh
@@ -3,11 +3,10 @@
 #-------------------------------------------------
 
 
-INSTALL="yay -S --noconfirm"
+INSTALL="yay -S --noconfirm --needed"
 COMPONENTS_PATH="./components"
 CONFIG_FILE_PATH="./alis-server.config"
 LOG_FILE="./alis-server.log"
-ADDITIONAL_CONFIG_FILES_DIR="files-to-deploy/additional-config-files"
 
 
 # Prevent sudo timeout
@@ -33,20 +32,31 @@ then
   exit 1
 fi
 
+HOMEDIR_DOTFILES_SOURCE="files-to-deploy/config-files/homedir"
+USER_HOMEDIR_DOTFILES_DESTINATION="/home/${username}"
+ROOT_HOMEDIR_DOTFILES_DESTINATION="/root"
+SCRIPTS_DIR="files-to-deploy/scripts"
+
+
 
 # Execute a component file
-# The name of the component (without the ending ".sh")
+# $1: The name of the component (without the ending ".sh")
 function install_component (){
-  echo "========================================" 2>&1 | tee -a ${LOG_FILE}
   echo "========================================" 2>&1 | tee -a ${LOG_FILE}
   echo "Starting installation of component $1" 2>&1 | tee -a ${LOG_FILE}
   echo "========================================" 2>&1 | tee -a ${LOG_FILE}
-  echo "========================================" 2>&1 | tee -a ${LOG_FILE}
-  source "${COMPONENTS_PATH}/$1.sh" 2>&1 | tee -a ${LOG_FILE}
-  echo "========================================" 2>&1 | tee -a ${LOG_FILE}
+  # Case components/core/core.sh
+  if [ -f "${COMPONENTS_PATH}/$1/$1.sh" ]; then
+      source "${COMPONENTS_PATH}/$1/$1.sh" 2>&1 | tee -a ${LOG_FILE}
+  # Case components/core.sh
+  elif [ -f "${COMPONENTS_PATH}/$1.sh" ]; then
+    source "${COMPONENTS_PATH}/$1.sh" 2>&1 | tee -a ${LOG_FILE}
+  else
+    echo "Error: Component $1 not found"
+  fi
+  echo "" 2>&1 | tee -a ${LOG_FILE}
   echo "========================================" 2>&1 | tee -a ${LOG_FILE}
   echo "Finished installing component $1" 2>&1 | tee -a ${LOG_FILE}
-  echo "========================================" 2>&1 | tee -a ${LOG_FILE}
   echo "========================================" 2>&1 | tee -a ${LOG_FILE}
   echo "" 2>&1 | tee -a ${LOG_FILE}
 }
@@ -68,25 +78,43 @@ function enable_networking (){
 }
 
 
+source ./global-variables.sh
+source ./common-functions.sh
+
+# Create directory for temporary files during docker compose files build
+  mkdir -p ${TEMP_DOCKER_COMPOSE_PARTS_DIR}
 
 enable_networking
-source common-functions.sh
-install_component networking
 install_component aur-helper
-install_component common-dependencies
-install_component vpn
+install_component networking
 install_component utils
 install_component security
 install_component core
 install_component file-manager
 install_component shell-and-term-related
 install_component ssh
-install_component transmission
-install_component web-server
-install_component kodi
-install_component sonarr-radarr
-install_component hyperion
-${COMPONENTS_PATH}/link-files.sh ${username}
+####
+install_component vpn-client
+install_component reverse-proxy
+install_component dns-updater
+install_component torrenting-client
+install_component media-center
+install_component media-center-2
+install_component home-assistant
+install_component system-monitoring
+install_component pvr
+install_component bookmarks-manager
+install_component finances-management
+install_component pihole
+install_component task-manager
+install_component file-server
+install_component file-synchronization
+install_component docker
+
+
+sudo_create_link ${SCRIPTS_DIR}/startup /bin
+sudo_create_link ${SCRIPTS_DIR}/manage-disks /bin
+
 
 sync
 sudo reboot
